move code to fusedlocation class
